Add different methods to select the place clicked on map
A number of approaches were tried. Most were attempting to work out the
most appropriate location to serve to the user based on zoom, either by
using the typ of each of the locations, or the bounding box relative to
the users view. However it was difficult to get it consistently right,
and would almost certainly be annoying for the user.

As such, a simple method of picking the first result, which is the most
specific, has been chosen.