Added search box
Used auto complete component to reduce returning multiple searches e.g.
is user searched for pubs will return multiple, which uses more API
resources, and thus more cost.

Also had to update the webpack css name rules, so that when overriding
the container for the search results it maintained the basic rule name,
and did not include the hash.